Correcao de erros, criacao e correcao do PokemonService, criacao da classe PokemonElementId

+++ src/main/java/com/ips/tpsi/pokemonwebapp/repository/ElementRepository.java
@@ -1,9 +1,10 @@
 package com.ips.tpsi.pokemonwebapp.repository;
+
 import com.ips.tpsi.pokemonwebapp.entity.Element;
-import org.springframework.data.jpa.repository.JpaRepository; 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ElementRepository
-	extends JpaRepository<Element, Long> { 
+public interface ElementRepository extends JpaRepository<Element, Integer> {
+
 }
