+++ src/main/java/com/ips/tpsi/pokemonwebapp/repository/PokemonElementRepository.java
@@ -1,9 +1,10 @@
 package com.ips.tpsi.pokemonwebapp.repository;
-import com.ips.tpsi.pokemonwebapp.entity.PokemonElement; 
-import org.springframework.data.jpa.repository.JpaRepository; 
+
+import com.ips.tpsi.pokemonwebapp.entity.PokemonElement;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface PokemonElementRepository
-	extends JpaRepository<PokemonElement, Long> { 
+		extends JpaRepository<PokemonElement, Integer> {
 }
